Assuming the working GNU assembler with --with-gnu-as
When configuring GCC with

--target=TARGET

to build a cross compiler to reproduce a compiler bug, as and collect have

ORIGINAL_AS_FOR_TARGET=""

As the result, many target features are disabled which makes it almost
impossible to reproduce the bug.  Without assembler, the GCC build won't
finish and the run-time libraries won't build.  But the GCC itself can
be built.  It is very useful for debugging GCC with a cross compiler.
Enable GNU assembler features for --with-gnu-as if assembler isn't
available.

	PR target/113684
	* acinclude.m4 (gcc_GAS_CHECK_FEATURE): Assuming the working GNU
	assembler for --with-gnu-as if assembler isn't available.
	* configure: Regenerated.