libstdc++: Fix -Wdeprecated warning about implicit capture of 'this'
Tested x86_64-linux. Pushed to trunk.

-- >8 --

In C++20 it's deprecated for a [=] lambda capture to capture the 'this'
pointer. Using resize_and_overwrite with a lambda seems like overkill to
write three chars to the string anyway. Just resize the string and
overwrite the end of it directly.

libstdc++-v3/ChangeLog:

	* include/experimental/internet (network_v4::to_string()):
	Remove lambda and use of resize_and_overwrite.