libstdc++: Replace padding bits with a bit-field in __format::_Spec
I plan to push this to trunk soon.

CC HP for visibility of the change affecting cris-elf. In practice it
shouldn't make any difference to any sensible code. It only affects
C++20 mode (and later), and only changes the size of std::formatter
objects which are typically only created by the library headers
themselves, and only on the stack (when using std::format and other new
C++20 APIs related to it).

-- >8 --

This ensures that the unused bits will be zero-initialized reliably, and
so can be used later by assigning them values in formatter
specializations. For example, formatters for std::chrono will need to
use an extra bit for a boolean to optimize the conversions between
locale encodings and UTF-8.

This will result in an ABI change for targets that use 1-byte alignment
for all integral types, e.g. cris-elf. We can't do that once C++20
support is non-experimental and ABI stable, so do it now before GCC 14
is released.

libstdc++-v3/ChangeLog:

	* include/std/format (__format::_Spec::_M_reserved): Define a
	new bit-field member in place of padding bits.