gcc/configure: Re-introduce INSTALL_INFO
BUILD_INFO is currently a byproduct of checking makeinfo
presence/version.  INSTALL_INFO used to be defined similarly, but was
removed in 2000 (!) by commit 17db658241d18cf6db59d31bc2d6eac96e9257df
(svn r38141).

In order to save build time, our CI overrides BUILD_INFO="", which
works when invoking 'make all' but not for 'make install' in case some
info files need an update.

I noticed this when testing a patch posted on the gcc-patches list,
leading to an error at 'make install' time after updating tm.texi (the
build reported 'new text' in tm.texi and stopped).  This is because
'install' depends on 'install-info', which depends on
$(DESTDIR)$(infodir)/gccint.info (among others).

This patch replaces the 'install-info' dependency in 'install' with
$(INSTALL_INFO), thus enabling to skip this step.

2024-02-01  Christophe Lyon  <christophe.lyon@linaro.org>

	gcc/
	* Makefile.in: Add INSTALL_INFO.
	* configure.ac: Add INSTALL_INFO.
	* configure: Regenerate.