c++/modules: anon union member of as-base class [PR112580]
Tested on x86_64-pc-linux-gnu, does this look OK for trunk?

-- >8 --

Here when streaming in the fields of the as-base version of
_Formatting_scanner<int> we end up clobbering ANON_AGGR_TYPE_FIELD
of the anonymous union type, since it turns out this type is shared
between the original FIELD_DECL and the as-base FIELD_DECL copy (copied
during layout_class_type).  ANON_AGGR_TYPE_FIELD first gets properly set
to the original FIELD_DECL when streaming in the canonical definition of
_Formatting_scanner<int>, and then gets overwritten to the as-base
FIELD_DECL when streaming in the the as-base definition.  This leads to
lookup_anon_field giving the wrong answer when resolving the _M_values
use.

This patch makes us avoid overwriting ANON_AGGR_TYPE_FIELD when streaming
in an as-base class definition, it should already be properly set at that
point.

	PR c++/112580

gcc/cp/ChangeLog:

	* module.cc (trees_in::read_class_def): When streaming in
	an anonymous union field of an as-base class, don't clobber
	ANON_AGGR_TYPE_FIELD.

gcc/testsuite/ChangeLog:

	* g++.dg/modules/anon-3_a.H: New test.
	* g++.dg/modules/anon-3_b.C: New test.