libdecnumber: fixed multiple potential access-out-of bounds errors by moving range conditions before reads.
Multiple `for` loops across `libdecnumber` contain boolean expressions where memory is accessed prior to checking if the pointer is still within a valid range, which can lead to out-of-bounds reads.

This patch moves the range conditions to appear before the memory accesses in each conjunction so that these expressions short-circuit instead of performing an invalid read.

libdecnumber/ChangeLog
       * In each `for` loop and `if` statement, all boolean expressions of the form `*ptr == value && in_range(ptr)` have been changed to `in_range(ptr) && *ptr == value`.

Bootstrapped on x86_64-pc-linux-gnu with no regressions.