middle-end/109559 - warning in system header not suppressed
set_inlining_locations looks at a possible macro expansion location
when the location is in a system header but it fails to update its
counter when there's no macro involved.  The following fixes that.

Bootstrapped and tested on x86_64-unknown-linux-gnu.

This doesn't fix the observed diagnostic in the PR which I think is
now given by design since we diagnose inlined code from a system
header into a function not in a system header.  But I think it still
fixes a bug.

The whole set_inlining_locations is a bit pointless since all that
matters should be the location of the call (and its system header
status) then.  I'll also note that -Wno-system-headers doesn't
help and we don't have any flag to disable diagnosing
inlined-from-system-header code either.

Unfortunately this is all from changes done by Martin Sebor so
it's difficult to tell the true intention.  The code in
set_inlining_locations doesn't really do what it documents
but fixing (I'll attach the "failed" patch in the PR) will
break testcases that test we diagnose inline copies.

Anyway - OK for the change below where I don't have any testcase.

Thanks,
Richard.

	PR middle-end/109559
	* tree-diagnostic.cc (set_inlining_locations): Always
	increment nsyslocs when loc is in a system header.