c++: avoid name lookup during mangling
Bootstrapped and regtested on x86_64-pc-linux-gnu, does this look
reasonable?

-- >8 --

It turns out that with modules we can call mangle_decl recursively,
which is a problem because the global mangling state isn't recursion
aware.  The recursion happens from write_closure_type_name, which calls
lambda function, which performs name lookup, which can trigger lazy
loading, which can call maybe_clone_body for a newly loaded function,
which can inspect DECL_ASSEMBLER_NAME, which enters mangling.  This was
observed when using fmtlib as a module with trunk, and leads to a bogus
"mangling conflicts with a previous mangle error" followed by an ICE
from cdtor_comdat_group due to a mangling mismatch.

I considered making our mangling state recursion-aware, but it seems
this lambda_function call is the only source of name lookup during
mangling, and in turn the only source of potential recursion with
modules so perhaps it's better to just address that.  To that end, this
patch adds a new field to LAMBDA_EXPR holding the op() of the closure
type, so that lambda_function can just inspect this field rather than
having to perform name lookup.  With this patch fmtlib can be
successfully imported as a module (with a few minor source tweaks).

In passing this patch adds streaming of LAMBDA_EXPR_REGEN_INFO which I
noticed is missing.

gcc/cp/ChangeLog:

	* cp-tree.h (LAMBDA_EXPR_FUNCTION): Define.
	(tree_lambda_expr::function): New member.
	* lambda.cc (lambda_function): Use LAMBDA_EXPR_FUNCTION instead
	of performing name lookup.
	* module.cc (trees_out::core_vals): Stream LAMBDA_EXPR_REGEN_INFO
	and LAMBDA_EXPR_FUNCTION.
	(trees_in::core_vals): Likewise.
	* parser.cc (cp_parser_lambda_declarator_opt): Set
	LAMBDA_EXPR_FUNCTION.
	* pt.cc (tsubst_lambda_expr): Likewise.