c++: Support lambdas attached to FIELD_DECLs in modules [PR111710]
Bootstrapped and regtested on x86_64-pc-linux-gnu, OK for trunk?

Or should I re-introduce the tree checking and just add checks for the
new kinds of declarations that could be have keyed decls?

-- >8 --

The fix for PR107398 weakened the restrictions that lambdas must belong
to namespace scope. However this was not sufficient: we also need to
allow lambdas keyed to FIELD_DECLs or PARM_DECLs.

Because this now requires 'DECL_MODULE_KEYED_DECLS_P' to be checked on a
fairly large number of different kinds of DECLs, and that in general
it's safe to just get 'false' as a result of a check on an unexpected
DECL type, this patch also removes the tree checking from the accessor.

gcc/cp/ChangeLog:

	* cp-tree.h (DECL_MODULE_KEYED_DECLS_P): Remove tree checking.
	(struct lang_decl_base): Update comments and fix whitespace.
	* module.cc (trees_out::lang_decl_bools): Always write
	module_keyed_decls_p flag...
	(trees_in::lang_decl_bools): ...and always read it.
	(trees_out::decl_value): Also handle keyed decls on decls other
	than VAR_DECL or FUNCTION_DECL.
	(trees_in::decl_value): Likewise.
	(trees_out::get_merge_kind): Likewise.
	(maybe_key_decl): Also handle lambdas attached to FIELD_DECLs
	and PARM_DECLS.
	(trees_out::key_mergeable): Likewise.
	(trees_in::key_mergeable): Support keyed decls in a TYPE_DECL
	container.

gcc/testsuite/ChangeLog:

	* g++.dg/modules/lambda-7_a.C: New test.
	* g++.dg/modules/lambda-7_b.C: New test.

Signed-off-by: Nathaniel Shead <nathanieloshead@gmail.com>