aarch64: Reword error message for mismatch guard size and probing interval [PR90155]
The error message is not clear what options are being taked about when it says the values
need to match; plus there is a wrong quotation dealing with the diagnostic.
So this changes the error message to be exactly talking about the param options that
are being taked about and now with the options, it needs the quoting.

OK? Built and tested for aarch64-linux-gnu.

gcc/ChangeLog:

	* config/aarch64/aarch64.cc (aarch64_override_options_internal): Fix
	error message for  mismatch guard size and probing interval.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>