OpenMP/C++: Fix (first)private clause with member variables [PR110347] [was: [RFA/RFC] C++/OpenMP: Supporting (first)private for member variables [PR110347] - or VALUE_EXPR and gimplify]
Hi,

your suggestion almost did the trick, but caused regressions with
lambda closures in target regions.

Jakub Jelinek wrote:
> Ah, and the reason why it doesn't work on target is that it has the
> everything is mapped assumption:
>    if ((ctx->region_type & ORT_TARGET) != 0)
>      {
>        if (ctx->region_type & ORT_ACC)
>          /* For OpenACC, as remarked above, defer expansion.  */
>          shared = false;
>        else
>          shared = true;
>
>        ret = lang_hooks.decls.omp_disregard_value_expr (decl, shared);
>
> Perhaps shared = true; should be shared = (flags & GOVD_MAPPED) != 0;
> now that we have private/firstprivate clauses on target?

Hence, I now use this code, but also pass a flag to distinguish target
regions (→ map) from shared usage, assuming that it is needed for the
latter (otherwise, there wouldn't be that code).

The issue only showed up for a compile-only testcase, which I have now
turned into a run-time testcase.
In order to do so, I had to fix a bogus test for is mapped (or at least
I think it is bogus) - and for sure it didn't handle shared memory.

I also modified it such that it iterates over devices. Changes to the
dump: the 'device' clause had to be added (3x) and for the long line:
'this' and 'iptr' swapped the order and 'map(from:mapped)' became
'firstprivate(mapped)' due to my changes.
I appended a patch which only shows the test-case differences as "git
diff" contains all lines as I move it to libgomp/.

Comments, remarks, suggestions?

Tobias

OpenMP/C++: Fix (first)private clause with member variables [PR110347]

OpenMP permits '(first)private' for C++ member variables, which GCC handles
by tagging those by DECL_OMP_PRIVATIZED_MEMBER, adding a temporary VAR_DECL
and DECL_VALUE_EXPR pointing to the 'this->member_var' in the C++ front end.

The idea is that in omp-low.cc, the DECL_VALUE_EXPR is used before the
region (for 'firstprivate'; ignored for 'private') while in the region,
the DECL itself is used.

In gimplify, the value expansion is suppressed and deferred if the
  lang_hooks.decls.omp_disregard_value_expr (decl, shared)
returns true - which is never the case if 'shared' is true. In OpenMP 4.5,
only 'map' and 'use_device_ptr' was permitted for the 'target' directive.
And when OpenMP 5.0's 'private'/'firstprivate' clauses was added, the
update that 'shared' is only true for 'map' was missed.

However, just enabling it for all '!shared' will cause issues with
Lambda closures ("__closure->this->...") for which also a DECL_VALUE_EXPR
exists but that is not related to DECL_OMP_PRIVATIZED_MEMBER. Solution:
Update the lang hook to take a Boolean argument, indicating whether it
is called for a target region or not.

2024-02-16  Tobias Burnus  <tburnus@baylibre.com>
	    Jakub Jelinek  <jakub@redhat.com>

	PR c++/110347

gcc/cp/ChangeLog:

	* cp-gimplify.cc (cxx_omp_disregard_value_expr): Add new
	Boolean argument and use it.
	* cp-tree.h (cxx_omp_disregard_value_expr): Update prototype.

gcc/fortran/ChangeLog:

	* trans-openmp.cc (gfc_omp_disregard_value_expr): Add
	unused Boolean argument.
	* trans.h (gfc_omp_disregard_value_expr): Update
	prototype.

gcc/ChangeLog:

	* gimplify.cc (omp_notice_variable): Update call to
	lang_hooks.decls.omp_disregard_value_expr.
	(omp_notice_variable): Likewise; fix 'shared' arg for
	(first)private in target regions.
	* hooks.cc (hook_bool_tree_bool_bool_false): New.
	* hooks.h (hook_bool_tree_bool_bool_false): New.
	* langhooks-def.h (LANG_HOOKS_OMP_DISREGARD_VALUE_EXPR):
	Use it.
	* langhooks.h (struct lang_hooks_for_decls): Add second
	Boolean argument.
	* omp-low.cc (omp_member_access_dummy_var): Update
	lang_hooks.decls.omp_disregard_value_expr call.

libgomp/ChangeLog:

	* testsuite/libgomp.c++/target-lambda-3.C: Moved from
	gcc/testsuite/g++.dg/gomp/ and fixed is-mapped handling.
	* testsuite/libgomp.c++/firstprivate-c++-1.C: New test.
	* testsuite/libgomp.c++/firstprivate-c++-2.C: New test.
	* testsuite/libgomp.c++/private-c++-1.C: New test.
	* testsuite/libgomp.c++/private-c++-2.C: New test.
	* testsuite/libgomp.c++/use_device_ptr-c++-1.C: New test.

gcc/testsuite/ChangeLog:

	* g++.dg/gomp/target-lambda-1.C: Moved to become a
	run-time test under testsuite/libgomp.c++.

Co-authored-by: Jakub Jelinek <jakub@redhat.com>

 gcc/cp/cp-gimplify.cc                              |   7 +-
 gcc/cp/cp-tree.h                                   |   2 +-
 gcc/fortran/trans-openmp.cc                        |   2 +-
 gcc/fortran/trans.h                                |   2 +-
 gcc/gimplify.cc                                    |  12 +-
 gcc/hooks.cc                                       |   6 +
 gcc/hooks.h                                        |   1 +
 gcc/langhooks-def.h                                |   2 +-
 gcc/langhooks.h                                    |   5 +-
 gcc/omp-low.cc                                     |   2 +-
 gcc/testsuite/g++.dg/gomp/target-lambda-1.C        |  94 -------
 libgomp/testsuite/libgomp.c++/firstprivate-c++-1.C | 305 +++++++++++++++++++++
 libgomp/testsuite/libgomp.c++/firstprivate-c++-2.C | 125 +++++++++
 libgomp/testsuite/libgomp.c++/private-c++-1.C      | 247 +++++++++++++++++
 libgomp/testsuite/libgomp.c++/private-c++-2.C      | 117 ++++++++
 libgomp/testsuite/libgomp.c++/target-lambda-3.C    | 104 +++++++
 .../testsuite/libgomp.c++/use_device_ptr-c++-1.C   | 125 +++++++++
 17 files changed, 1048 insertions(+), 110 deletions(-)